- Began working on volume transform library (Volt) - Renamed RawImage to Bitmap (Actually was an existing unimplemented class, was intended to be Bitmap to begin with) - Fixed tonemapping exposure parameter interface - Fixed bug with const factor for scatter and surface integrators being different - Changed VectorXu to unsigned int rather than size_t data type - Minor bugfixes/refactors

+++ Source/VoxLib/Core/Geometry/Vector.h
@@ -635,9 +635,9 @@ namespace vox
 	typedef Vector<float,2> Vector2f;
 	typedef Vector<float,3> Vector3f;
 	typedef Vector<float,4> Vector4f;
-	typedef Vector<size_t,2> Vector2u;
-	typedef Vector<size_t,3> Vector3u;
-	typedef Vector<size_t,4> Vector4u;
+	typedef Vector<unsigned int,2> Vector2u;
+	typedef Vector<unsigned int,3> Vector3u;
+	typedef Vector<unsigned int,4> Vector4u;
 }
 
 // End Definition
